requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------
ultralytics==8.0.232
gitpython>=3.1.30
matplotlib>=3.3
numpy
opencv-python>=4.6.0
Pillow~=9.5.0
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy
thop>=0.1.1  # FLOPs computation
# torch==1.9.0+cu111  # see https://pytorch.org/get-started/locally (recommended)
# torchvision==0.10.0+cu111
tqdm>=4.64.0
# protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012
timm

# Logging ---------------------------------------------------------------------
wandb
# tensorboard>=2.4.1
# clearml>=1.2.0
# comet

# Plotting --------------------------------------------------------------------
pandas>=1.1.4
seaborn

# Export ----------------------------------------------------------------------
# coremltools>=6.0  # CoreML export
onnx~=1.12.0  # ONNX export
onnx-simplifier>=0.4.1  # ONNX simplifier
# nvidia-pyindex  # TensorRT export
# nvidia-tensorrt  # TensorRT export
# scikit-learn<=1.1.2  # CoreML quantization
# tensorflow>=2.4.0  # TF exports (-cpu, -aarch64, -macos)
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev>=2023.0  # OpenVINO export

# Deploy ----------------------------------------------------------------------
setuptools>=65.5.1 # Snyk vulnerability fix
# tritonclient[all]~=2.24.0

# Extras ----------------------------------------------------------------------
# ipython  # interactive notebook
# mss  # screenshots
# albumentations>=1.0.3
# pycocotools>=2.0.6  # COCO mAP
